refactor(state): extract CompileCoordinator from Model
Move the compile lifecycle — processSource (source materialization + the #49
fetch-race fix), checkSyntax, render (preview/full + supersession guards, the
#56 revision stale-drop, dimension retry), rawStreamsCallback, and the
sequence counters — into a CompileCoordinator behind the shared ServiceContext.
Model.render()/checkSyntax() become delegations; the source/file-op paths call
this.compile.processSource().

The source-revision bump stays in Model.setState (it fires on source changes
that originate in file-ops too); the coordinator reads it via the context. The
'state' event and persistence funnel are untouched.

Behavior-identical: the unchanged model/source-fetch-race/checkingSyntax
oracles still pass, and an adversarial line-by-line review found no required
fixes. Model drops from ~690 to 423 lines.

Refs #59
